Reorganize service
