+++ packages/nuxt-locokit/src/runtime/pages/admin/groups/CreateGroup/CreateGroup.vue
@@ -139,10 +139,10 @@ import { FormGeneric } from '@locokit/designsystem'
 import { Field } from 'vee-validate'
 import { ref } from 'vue'
 import { ROUTES_NAMES } from '../../../../locokit-paths'
-import { createGroup } from '../../../../services/group'
+import { createGroup } from '../../../../services/core/group'
 import { Policy, Workspace } from '../../../../interfaces/toMigrate'
-import { findWorkspaces } from '../../../../services/workspace'
-import { findPolicies } from '../../../../services/policy'
+import { findWorkspaces } from '../../../../services/core/workspace'
+import { findPolicies } from '../../../../services/core/policy'
 import { useStoreGroups } from '../../../../stores/groups'
 import { useRouter } from '#imports'
