+++ packages/nuxt-locokit/src/runtime/pages/admin/users/index.vue
@@ -190,7 +190,7 @@ import { useI18n } from 'vue-i18n'
 import { COLUMN_TYPE } from '../../../helpers/filter'
 import { ROUTES_NAMES } from '../../../locokit-paths'
 import { useStoreUsers } from '../../../stores/users'
-import { searchUsers } from '../../../services/user'
+import { searchUsers } from '../../../services/core/user'
 import { ApiUser, Filter, PROFILE } from '../../../interfaces/toMigrate'
 import { ref } from '#imports'
