+++ packages/nuxt-locokit/src/runtime/services/core/usergroup.ts
@@ -1,6 +1,6 @@
 import { SERVICES } from '@locokit/definitions'
 import { UserGroupData } from '@locokit/sdk'
-import { sdkClient } from './api'
+import { sdkClient } from '../api'
 
 const ITEMS_PER_PAGE = 20
